[flickr] Add async version for flickcurl_photos_search_params() function
Unfortunately, flickcurl a blocking library, meaning that whenever a function
is invoked, it blocks until it returns.

This gives a bad user experience, as everything is blocked while function is
running.

To avoid this, an async version for flickcurl_photos_search_params() function
is provided, where it runs the original function in its own thread, and result
is sent through a callback.