[shoutcast] Add search term to id
When browsing a genre, resulting media id is formed by genre and station id.

When searching, as no genre is specified, we were using an empty genre.

For future development of metadata() function, add the search term as it were a
genre.

In order to distinguish in future the search term from a real genre, prefix the
term with '?'.

Thus, a result id from "Pop" category would be "Pop/129322", while a result
from searching "Foo" would be "?Foo/129322".