[upnp] Use UPnP source name as title of root category
So far, titles for UPnP-medias were obtained while browsing through UPnP
sources. Problem is that doing it, root category would never have a title, as
it is not obtained from a UPnP browse.

The fix uses the name of original UPnP source as the title for root category.
Note that the name of the created Grilo source is the name of original UPnP
source prefixed with "UPnP - " string.