[shoutcast] Change id encoding in search()
So far, id was formed by concatenating Genre + StationID, except in results
from search, where was encoded as "?" + Search text + StationID.

As it can be seen, two different ways of encoding id.

But 2nd-way encoding has a problem: as SHOUTCast does not provides an API to
query metadata, it is resolved in plugin as repeating the same operation that
led to result, and then getting it from the results. Thus, in first case it
browses the same Genre and then pick the station id, while in the second case
it searches the same string and then pick the result (that is why the "?" is
needed).

The problem is that it is possible than when repeating the searching, the
element is not retrieved. Reason is that could be in the first search a
different limit was applied, so in the second search limit is not the same and
thus different results are retrieved.

To avoid this problem and in order to keep a consistent ID, this patch gets rid
of second approach and always uses the first one: id is encoded as "Genre +
StationID.