flickr: Set up personal sources
Personal sources are intended to access public and private content of a
specific user (the one who has authorized the plugin to access its content).

This personal source appears when user creates a configuration for Flickr
containing a token.

In personal sources, first-level content are the user's photosets, while the
second level are the photos in that photoset.

Also, searching is restricted to personal content; if user wants to search in
all Flickr, it should be done in the public source (the one with no token at
all).