tracker: Add Tracker plugin
Plugin that searches multimedia content using Tracker.

This first version implements the query function. A query must be a sparql
query that Tracker can understand.

In order to make possible to map Tracker results to Grilo, all columns must be
named with the name of metadata key, and must include a column telling the type
of result returning. More information in grilo_tracker_source_query() function.

Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>