tracker: Get rid of supported_operations()
It is not needed to redefine this function. When providing a
implementation for an operation, core automatically detects this and
assume source is able to do that operation.

Only in the case where though providing an implementation, source can
still not be able to do it, we need to provide our custom
supported_operations().

Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>