tracker: add per device mode support
This adds the "per-device-source" option to the plugin, which allows
to browse/search/etc... content on a per data source (in sense of
Tracker) basis. The main application to this is, for example, to
browse content from a USB key without taking care of files coming from
the local filesystems.

When the per device mode is activated, the plugin looks for available
data sources within the Tracker database and creates a media source
per data source and monitor changes on them.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@linux.intel.com>
Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>