filesystem: added cancellation support
For search, this is done easily by passing a GCancellable to GIO calls.
Since browse does not use GIO's async API, the only useful thing to do
(assuming an operation cannot be cancelled from another thread), is to check if
the operation has been cancelled at the start of browse_emit_idle(), which is
when the browse operation continues execution in the only cases where it had
delayed some of its execution.

Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>