local-metadata: added a new local source
For the moment it only implements thumbnails for images and videos. It does not
create any thumbnail but uses GIO's G_FILE_ATTRIBUTE_THUMBNAIL_PATH to find out
if there is one existing.

There's a plan to support the Media Art Storage Spec as well, but it is not
implemented yet.

Signed-off-by: Iago Toral Quiroga <itoral@igalia.com>