tracker: metadata should not resolve media coming from tracker
This is mainly an optimization: we do not want tracker resolving keys for
medias coming from tracker.

The rationale is that we have metadata() precisily to do it: invoking resolve()
after invoking metadata() for Tracker media is useless, as if the key could not
be resolved with metadata() it will not be resolved with resolve() (both
functions are sharing the same code).

This fixes https://bugzilla.gnome.org/show_bug.cgi?id=645523

Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>