youtube: Do not preload categories
Get categories and feeds only when it is required, instead of doing when
loading the plugin.

This will prevent getting information that possible will not be used (if user
do not perform a browse).

As consequence of this, childcount is only computed when the user really
browses through the category.

This fixes https://bugzilla.gnome.org/show_bug.cgi?id=650679

Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>