upnp: treat operation failure as something expected
When using gupnp_service_proxy_end_action() ignore the returned result, and
just check if the required data has any value or not.

Reason is that with some versions, not having elements matching the required
search text means getting an error in that operation, while in others it
doesn't mean it, but just an empty result.

As we want to treat not having elements as a typical case, better do not print
the error as a warning, and instead notify that there are no results.

This fixes https://bugzilla.gnome.org/show_bug.cgi?id=653756

Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>