local-metadata: Don't leave unfinished calls
It was possible to hang an application that would have been
waiting for metadata because of us not calling the callback
in some cases.

We must make sure to either have async calls ongoing from which
the callback will be called, or call the callback straight away.

Fixes grilo-test-ui hanging when browsing large UPnP containers
full of videos to parse.

https://bugzilla.gnome.org/show_bug.cgi?id=673936