Link upnp plugin to libgssdp
The upnp plugin uses gssdp but doesn't link to it.
This used to work since gupnp leaked linker flags.
Since gupnp 0.19.0 the pkg-config files has been
fixed, so this means grilo-plugins needs to be fixed
to link to everything it needs directly and not
relying on anything else pulling in linker flags.