flickr: Deal with photos only
While Flickr can store videos, API is mainly designed to deal with photos.
Actually, when we get a video from Flickr, its current URL do not point to the
video itself, but rather to a thumbnail capture.

As most people identify Flickr with photo storage service, let's focus in
photos and do not return videos.