upnp: Fix a race condition when creating a new source
When creating a new source, before registering it, it is checked if the device
supports search capability, so the source is registered with or without search
operation, accordingly.

Checking the capability is an asynchronous operation, and can happen that while
that operation is running, the device disappears. This means that when the
search capability operation finishes and call the proper callback, it must be
checked if the device was removed or not. If it was, then drop the source, else
register it.