upnp: Add filter by type in browse() operation
UPnP/DLNA protocol does not allow to apply filters in the browse operation.

So the way to implement it is to rely on the search() operation: browsing in
container "foo" is like searching all the elements with @parentID=\"foo\". And
thus we can apply here the filter by type (allowing also containers).

This means that filter by type is only available in case the source supports
search() operation.