youtube: Make sure structures are properly initialized
Some of the used structures are initialized in the first browse/search
operation. This because assume that a resolve() must be done always
after a previous search/browse.

But it can happen that the element to be resolved comes from a
browse/search in a previous session. So resolve() can be the first
operation and the structures are not properly initialized.

This commit fixes this use case.