lastfm-albumart: Do not return default image
When Last.fm does not recognize an album, it returns a default image to be used
as the thumbnail. This prevents other sources to take the chance to try to
provide a real thumbnail.

Do not return a default thumbnail if the source can not solve it, and let other
sources to try it.