plugins: Fix possible crash when registering sources
Calls to grl_registry_register_source() can lead to the source
being unref'ed (for example, a video player not interested in audio-
only sources would do that in "source-added"), so we must be careful
when using the source after registration.

Both the dmap and podcasts plugins used the source after registration,
so we must make sure they don't use freed memory.