magnatune: Adding support to download database
All database and crc checks are made when user interact with the plugin.
If there is no database and no internet connection, plugin_init fails.
When current database has 7 days old the plugin starts to check for a
newer database.
To avoid checking crc to *every* user request a time interval of 12
hours is used.

https://bugzilla.gnome.org/show_bug.cgi?id=698523