raitv: Fix crasher on 64-bit platforms
gsize on 64-bit platforms is a 64-bit integer. Trying to fit a 64-bit
integer in the 32-bit integer we were passed trashed the pointer of the
previous return parameter, in this case the buffer's address.

Pass a pointer to correctly sized integer to avoid this memory
corruption and cast it later as needed.

https://bugzilla.gnome.org/show_bug.cgi?id=700297