upnp: Use container "0" when browsing with filters
Browsing when filters are involved is implemented through "Search" capability,
because it allows to specify filters.

As "Search" is recursive, and most of servers support it only in the root
container, use the root container when searching instead of using the current
container, in order to be able to use those filters.