dleyna: Plugin to access DLNA contents using the dLeyna DBus service
This plugin uses the dleyna-server service to interact with DMS (DLNA
Media Server, eg. Rygel) through an high level DBus API and it is meant
to deprecate the UPnP plugin.

It implements all the functionalities exposed by GrlSource: resolve(),
browse(), search(), query(), store(), store_metadata(), remove(),
cancel() and change notifications.

This adds a soft-dependency on the deyna-renderer DBus service: if the
service is not activatable no DLNA server sources will be shown.

The required GLib version has been bumped to 2.36 to use GTask.

https://bugzilla.gnome.org/show_bug.cgi?id=707346

Signed-off-by: Emanuele Aina <emanuele.aina@collabora.com>