lua-factory: Grilo operations to Lua sources
This is the glue code between Grilo and the Lua sources. The Lua stack
(lua_State) is used to pass the variables to the lua code.

Besides the Grilo operations, the lua source can use the grl_source_init
function to check if it can be successfully loaded or not.

https://bugzilla.gnome.org/show_bug.cgi?id=711243