fake-metadata: Remove Fake Metadata plugin
The purpose of this plugin was to be an example of how to write a plugin
providing metadata information to other plugins.

Given there are several plugins provides this information (see Last.fm or
Gravatar plugin), it doesn't make sense to keep this plugin, specially when the
data it was providing was totally a fake.

So let's remove it.