lua-factory: Assert when overwriting user_data with concurrent resolves
This will assert when multiple resolve calls are launched concurrently
for the same lua source.

We would set the operation spec for that state when calling resolve, but
as it is asynchronous, the state would get assigned multiple operation
specs, with the latter overwriting the earlier one.

This is easily reproduced by browsing in grilo-test-ui and select
different video items quickly, so one of the lua sources that implements
resolve() gets called twice with the same lua state.

The fix might be to have one lua state per call, or have a way to get
the operation spec for that particular resolve call.

https://bugzilla.gnome.org/show_bug.cgi?id=726563