tracker: Correctly set "title-from-filename"
Tracker usually provides us with a title that's derived from the
filename. Replicate its process to create a title from a filename
to check if the title is actually set from the filename.

This allows the local-metadata plugin to parse it to remove gunk, and
other plugins to override its throwaway title.

https://bugzilla.gnome.org/show_bug.cgi?id=740756