lua-factory: Fix may_resolve with NULL media
grl_lua_factory_source_may_resolve might be called with media being NULL
(via expand_operation_keys in grl-source.c for example). Currently this
case is not handled properly and may_resolve calls grl_data_has_key
which results in a critical warning because media is NULL.

This is one of the sources of some of the critical warnings when running
gnome-music.

https://bugzilla.gnome.org/show_bug.cgi?id=744353