youtube: Use GDataYouTubeQuery when building GData queries
Do not use the parent GDataQuery class. While this worked previously, it
is not a valid use of the API, as it prevents any overrides in
GDataYouTubeQuery from being used when building query APIs. This will
cause queries to break in the transition from Google’s v2 to v3 API,
since the GDataYouTubeQuery behaviour has to diverge from GDataQuery.

https://bugzilla.gnome.org/show_bug.cgi?id=748395