tracker: Do not advertise browse support for the Tracker extractor
The Tracker extractor data source (which appears when using the
per-device-source option) does not support browsing, only resolving.
Override the supported_operations vfunc to ensure that GRL_OP_BROWSE is
not advertised for that source.

This fixes clients which want to browse a single Tracker source (e.g. a
single removable drive) and are checking for browsability using
grl_source_supported_operations().

https://bugzilla.gnome.org/show_bug.cgi?id=748173