local-metadata: Allow tracking of multiple asynchronous resolve ops
Currently, only a single asynchronous operation is (potentially) used to
resolve local metadata. However, to fix the problem where the
thumbnail may be set to a non-existent URI, a second asynchronous
operation needs to happen in parallel. The current GrlSourceResolveSpec
closure cannot track this, so a new wrapper closure has to be added
which tracks the number of pending sub-operations and only calls the
GrlSourceResolveSpec’s callback once all are complete, or when a single
one errors.

https://bugzilla.gnome.org/show_bug.cgi?id=752057