local-metadata: Query for media art existence
Don’t set the media art on a GrlMedia if it doesn’t exist — libmediaart
unconditionally returns a path for where it would expect the media art
for a file to be, without checking that some art actually exists at that
path. That’s our job.

Asynchronously query for the file info to check it exists. Ignore
whether the media art is readable; that’s for the caller to handle when
they try to read it (and is consistent with the other thumbnail URIs
Grilo returns).

https://bugzilla.gnome.org/show_bug.cgi?id=752057