local-metadata: try per-track XDG thumbnails before per-album media art
Some XDG thumbnailers (including at least Tumbler and Totem's
thumbnailer) are able to extract embedded art from media files and cache
it in XDG_CACHE_HOME/thumbnails, just like they would for images and
videos. In some cases the thumbnailer might support embedded art even
when the libmediaart extractor (in my case tracker-extract) does not.

If this embedded art exists, it is likely to be a better match
for the track than the album art would be: for instance,
http://freemusicarchive.org/music/Nine_Inch_Nails/The_Slip/ has a
distinct piece of artwork embedded in each track, none of which
match the album cover. Accordingly, I'm trying the thumbnail before
falling back to libmediaart, not the other way round.

Using the thumbnailer's interpretation of the correct per-track
thumbnail also has the benefit that the track is guaranteed to look
the same in a Grilo application as it does in a general-purpose
file manager, whereas with libmediaart it is possible that the
thumbnailer and the libmediaart extractor might choose different
images in corner cases.

In principle, https://wiki.gnome.org/DraftSpecs/MediaArtStorageSpec
supports per-track artwork, which libmediaart could process and store;
the specification recommends treating embedded art as the highest
possible priority. However, in practice libmediaart does not
currently do this.

Bug: https://bugzilla.gnome.org/show_bug.cgi?id=752437