local-metadata: don't double-free cancellable
resolve_album_art() is "balanced" with respect to the cancellable's
refcount: it exits with 1 reference to the cancellable owned by
the operation_id, having arranged for that reference to be destroyed
by a GDestroyNotify when the operation_id finishes. There
is no need for resolve_album_art_cb() to release the same reference.

Bug: https://bugzilla.gnome.org/show_bug.cgi?id=752057