bliptv: Remove source
As stated in the blip.tv/shutdown/message.html
"Dear Blip User,

Blip was closed on Thursday, August 20, 2015.

As you may know, Blip was acquired by Maker Studios in September 2013
and the acquisition has allowed for additional, more expansive
direct-to-consumer tools and products for content creators across the
network.

If you are a creator and you have a YouTube channel, we encourage you
to apply to the Maker Gen network to take advantage of Maker’s tools
and services. Signup can be found here: https://max.makerstudios.com/.

If you are a viewer, we encourage you to check out the great content
from Maker talent at our new flagship site: http://maker.tv/

We thank you for being a great partner to Blip!

Blip Support"

This source does not work anymore.

https://bugzilla.gnome.org/show_bug.cgi?id=725031