tests: new function to reset test environment
With commit b4cf71a18de08825006ed7e7c3b8545e736b1839 we are using
temporary database for thetvdb tests. The goal of this patch is a way to
reset test environment by removing the database and re-loading the
plugin. If a test need a clean environment, it can call
test_reset_thetvdb ().

https://bugzilla.gnome.org/show_bug.cgi?id=748422