lua-factory: change grl.get_media_keys() into the static table
It's part of the work aimed at removing the OperationSpec from the
global scope.

grl_l_media_get_keys was replaced by grl_lua_library_push_grl_media

Although the GrlMedia is now a parameter for the lua source, Grilo
expects that the same GrlMedia object will be returned in the callback
for the Resolve operation [0]; For that reason we still keep track of
GrlMedia and merge it with the resulting GrlMedia from Lua source.

[0] see https://bugzilla.gnome.org/show_bug.cgi?id=760382#c3

https://bugzilla.gnome.org/show_bug.cgi?id=753141
Acked-by: Victor Toso <me@victortoso.com>