acoustid: including tests for the lua source
I'm including real fingerprint of four recordings in order to make this
test an easy way to improve acoustid in the future. For daily tests, the
net is mocked.

It is also important to note that acoustid needs a metadata-key
created by chromaprint which makes necessary to load grl-chromaprint
for this test.

https://bugzilla.gnome.org/show_bug.cgi?id=732879