lua-factory: use requested-keys as keys for lua api
We always provided the requested-keys to the lua source as an array with
the metadata-keys as values. That's not so interesting as the source
will need to walk in the array to check which keys were requested.

As from commit 2bfcff90d589d43351 we started introducing the
requested-keys as an argument, it would be good to use this moment and
improve it.

A table with the metadata-keys as key could be easily accessed in the
source

e.g
if requested_keys.artist then
  media.artist = my_artist
end

https://bugzilla.gnome.org/show_bug.cgi?id=732879