Revert "lua-factory: change grl.callback(), grl.fetch() and grl.unzip()"
In order to keep the previous API between Lua sources and Lua-Factory.

This reverts commit d809be39ed2bbd4c410be7aa527b36902362511a.

After applying this patch, it will break the lua sources due the API
change.  We are not passing callback as argument anymore. Lua sources
should use grl.callback().

This is not 100% revert as we are keeping a few changes to grl.fetch and
to grl.unzip:
* Both those functions are still using a lua function as callback which
  was a string before d809be39;
* Both this functions now can handle userdata as the last argument;
* For grl.fetch, the second argument is still an optional table for
  GrlNetWc;

https://bugzilla.gnome.org/show_bug.cgi?id=763046