Revert "lua-factory: change grl.get_media_keys() into the static table"
In order to keep the previous API between Lua sources and Lua-Factory.

This reverts commit 3e409b7727f803e2be4427c7544c679373a7d926.

This breaks the current plugins by changing the static table passed as
argument back grl.get_media_keys()

After applying this patch, it will break the lua sources due the API
change.  We are not passing an static table as GrlMedia anymore. Lua
sources should use grl.get_media_keys().

This is not 100% revert. We are keeping some fixes that was introduced
together with the api change for grilo-0.3 on grl.get_media_keys such as
using GrlMedia instead of specific types for audio, video, image and box
(which is now container)

https://bugzilla.gnome.org/show_bug.cgi?id=763046