Revert "lua-factory: change grl.get_options() into the static table"
In order to keep the previous API between Lua sources and Lua-Factory.

This reverts commit 2bfcff90d589d4335105a6423616c7de61cf4c71.

After applying this patch, it will break the lua sources due the API
change.  We are not passing an static table with all operation options
to the source anymore.  Lua sources should use grl.get_options() and
grl.get_requested_keys().

The main goal of reverted patch was to provide a table with all options
for the operation handler.

https://bugzilla.gnome.org/show_bug.cgi?id=763046