lua-factory: watchdog for lua-sources
This commit allow us to track broken sources such as:
* sources that never call for grl.callback()
* sources that might call grl.callback() after the operation is
  already finished

This is done by using a userdata in the top of the stack as a watchdog
for the current operation. We always push this userdata in
grl_lua_library_pcall function and it will be always freed by Lua's
garbage collection.

By attaching a finalize function into this userdata, we can check at
the time that lua_gc is called if the state of Operation is correct.

This idea was introduced by Morse in d809be39ed2bb so I modify those
functions to work as the watchdog we need instead of grl_l_callback
closure.

https://bugzilla.gnome.org/show_bug.cgi?id=763046