[0.2.x] tracker: Fix type of media when it has multiple types
To avoid a Video application receiving an audio file, when a file is
tagged in Tracker as having both types, and possibly shifting the
problem to another type if we were to have a simple "priority" list,
check what the application wants through the GrlTypeFilter it sets, and
prefer creating a video file if the app prefers videos.

This should stop totem receiving GrlMediaAudios when it only ever
expects GrlMediaVideo or containers.

https://bugzilla.gnome.org/show_bug.cgi?id=730028

Conflicts:
	src/tracker/grl-tracker-utils.c