lua-factory: support to cancel operation
Grilo has optional method to cancel operations which Lua-Factory had
not implemented so far, till now.

For each operation we create a GCancellabe that should be used in
async functions in Lua-Library such as grl.fetch and grl.unzip.

The grl_lua_operations_cancel_operation(), will cancel the operation,
remove the OperationSpec from its internals and free the related data.

https://bugzilla.gnome.org/show_bug.cgi?id=764077