lua-factory: Add GameFAQs source
To fetch metadata for games.

Now that grilo core doesn't force a particular sub-type of media
(video, audio or image), and given that the main user of this
source will be gnome-games, it makes sense to ship this helper
in grilo-plugins directly.

With help from Adrien Plazas <kekun.plazas@laposte.net>

https://bugzilla.gnome.org/show_bug.cgi?id=764602