lua-factory: Keep GrlNetWc saved as property
In order to make usage of throttling from GrlNetWc we need to keep it
stored to be used between different operation calls to the lua source.

In order to store this pointer, we make usage of source table which is
explained in 1074349f2f6ca3edc939313c7ff7d8ff2f5ae53b.
The full path for this new table and the GrlNetWc userdata is
* grl.__priv_state.properties
 - grl.__priv_state.properties.net_wc

In order to avoid leaks from these properties, the priv_state has now
a metattable with __gc set.

https://bugzilla.gnome.org/show_bug.cgi?id=769331