metrolyrics: fix html parser
Seems that in some lyrics a new <div></div> can be included with some
info. That was breaking the html parser.

Instead on relying on ending </div> for the lyric, let's use something
else that is present and not so common (<p class="writers") as ending
point for the lyrics and remove what is not interesting for us.

https://bugzilla.gnome.org/show_bug.cgi?id=770806