tracker: Support favorite read/write
Support GRL_METADATA_KEY_FAVOURITE reading & writing in the tracker
backend. The tracker ontology for this uses hasTag in combination with
the nao:predefined-tag-favorite property, which doesn't line up directly
with the boolean format Grilo uses. So either set the key with the given
property or delete the tag completely.

https://bugzilla.gnome.org/show_bug.cgi?id=774748