tracker: Remove per-device-source configuration
This is not entirely right, and could yield errors similar to commit
73c1aa1e7, where events go undetected because the deletion time of
properties is not guaranteed to be atomic, so there might be missed
events where files no longer qualify for certain notifications.

Given this is a configuration toggle meant for plugins/integrators,
it seems we can do without it.

Closes: #14