lua-factory: Fix possible crash when fetch ops are cancelled
When an operation was cancelled, we would go to free our operation data
straight away, including struct members shared with other operations.

As those could be on-going, and we know that the operations data members
that are private to our operation are valid, go check whether we are the
last operation, and only free the shared resources in that case.

https://bugzilla.gnome.org/show_bug.cgi?id=773789