travis: use Docker to perform build and test
Finally, we move to use containers in Travis CI to build and test Grilo.

It uses grilofw/grilo:core as base image.

If build goes fine, it will push grilofw/grilo:full container, which
contains Grilo with all the plugins. Users can get it to play with
Grilo.

It also pushes the test coverage statistics to Coveralls.io.

Build and test is done using autotools, as well as using Meson build,
both in parallel.