lua-factory: Sanitize code generation
Now that gperf is required to build this plugin, the generated
htmlentity.c should be removed from the source. Otherwise, it
will be used iff builddir == srcdir.

Add some further adjustments to make sure out-of-tree builds don't
break.

gperf 3.1 changes the type of the len argument for the generated
function to size_t, causing a conflict. Replace the prototype with
pragmas to make GCC ignore -Werror=missing-prototypes.

Tested with "make distcheck" and meson, with gcc and clang.

https://bugzilla.gnome.org/show_bug.cgi?id=778857