Fix build with latest gmime3
This has been broken for weeks....

Also, drop support for older gmime since we're early in the release
cycle and there's not really any compelling reason not to. Obviously
this means there must be no more grilo-plugins releases until the next
gmime release. If this is undesired then grilo-plugins could grow
conditional support for both versions of gmime, but it seem pointless.