tracker: Use standard Sparql1.1 syntax
The "AS var" syntax was non-standard, the Sparql1.1 spec defines this
syntax as "AS ?var", which Tracker 1.x has also accepted for a very long
time.

Tracker 2.x is dropping the non-standard syntax, so switch the Tracker
plugin to using the Sparql1.1 syntax, which shall work with Tracker 1.x
and 2.x.

https://bugzilla.gnome.org/show_bug.cgi?id=786602