lua-factory: fix bug for non-US locales
The steam-store lua plugin date formatter can only handle dates from the
US locale. To solve this issue, do two requests. One with the US locale
and one with the system locale. Take dates from the US locale (which can
be parsed) and all other information from the system locale (which is
localized for the user).

fixes #44