lua-factory: steam-store bug fix for non-US locale
The steam-store lua plugin date formatter can only handle dates from the
US locale. To solve this issue, do another request for the data with the
US locale if the request with the system locale contains a date that
cannot be parsed and use the date from the US locale.

fixes #44