theaudiobd: Add support for artist art
Before, we could only resolve thumbnail for albums. With this commit,
if only artist key is set, we can resolve the artist thumbnail as
well. That means we remove the album as a required metadata-key for
resolve operation and allow it to be optional.

The unit test is also updated to test the artist art retrieval.