tracker: Remove musicbrainz keys from the list of writable keys
MusicBrainz metadata have been added to Tracker ontology during the
2.3 development cycle. The initial idea was to extend nmm:MusicPiece
to add new MusicBrainz properties. This solution was not generic
enough (it did not allow to handle other online services). Finally, it
has been decided to introduce tracker:ExternalReference to handle
MusicBrainz (and any ony online service).

This plugin can only update a property of a nmm:MusicPiece (for
example its title). However, with the selected ontology, the
musicbrainz keys are not a property of a nmm:MusicPiece but an
external resource associated with a nmm:MusicPiece. Therefore,
GRL_METADATA_KEY_MB_RECORDING_ID and GRL_METADATA_KEY_MB_TRACK_ID keys
cannot be writable.

The next commit will update the mapping to use the correct ontology:
tracker:ExternalReference.