tracker-utils: use runtime check for ExternalReference queries
This patch allows the tracker plugin to be built in not so recent
versions of tracker library. This patch reverts the build time check
introduced in 7c29d5e and avoids regression within new queries by
doing a runtime check before setting up the mapping between tracker
and Grilo's metadata-keys.

Resolves: https://gitlab.gnome.org/GNOME/grilo-plugins/issues/55