tests: Add a test for lua-factory GrlRelatedKeys support
json test data are loaded from the file
grl-media-test-related-keys.json. It contains a key named
"related-keys" which is an array of related keys.

On the lua side, these "related-keys" are converted into indexed
arrays in order to create a GrlMedia with related keys.

On the other side, when the data are read from `json-glib.h`, the
related keys are retrieved by looping through "related-keys" array.