tracker3: Drop browse-filesystem configuration option
Besides allowing to browse folders as media containers, this option
makes change notifications work by enabling tracker:notify on nfo:Folder
type, so the grilo client (and everyone else, really ) may receive
TrackerNotifier notifications about folder additions/changes.

It's not in the hand of a client to do this, it ought not. Besides the
option seems completely unused according to codesearch.debian.net.
The Tracker plugin should not be intended for browsing as a filesystem
structure, and it is possible to create media containers specifically
for folders that do contain media. Drop this configuration option in
the tracker3 module.