tracker3: Drop request queue
For a large part, this work duplicates tracker's internal handling
(eg. updates are already guaranteed to be executed sequentially), or
even hinders it (eg. select queries are serialized even though they
can run in parallel).

This will also get a bit in the middle, as it makes sense to
minimize the time between a query happening and a
TrackerSparqlStatement being executed for that TrackerSparqlStatement
to be ready for reuse ASAP.

Drop this queue and let Tracker handle things itself.