Add option to disable bulding help
Help is meant to be a set of examples to application developers who
are interested in using grilo-plugins in their applications.

This patch addes a build option to not build and install those files
files (only one, using TMDB, but it installs it once per LINGUAS/* at
least. This is not needed in a sandboxed application like Flatpak

Closes: https://gitlab.gnome.org/GNOME/grilo-plugins/-/issues/13