tests: Adapt to tracker-test-sandbox utility changes
With the rename of the Tracker & Tracker Miners projects to TinySPARQL
and LocalSearch, this testing utility has been moved between projects
(from the SPARQL library, to the indexer that it's closely related to).

This brought up changes in the way it is found and invoked, this
utility is now available as a "localsearch test-sandbox" command, found
in $PATH. This will apply with TinySPARQL and LocalSearch >3.8.alpha.

Adapt the grilo testing framework to use either the old or the new
way, depending on what is found in the environment.